fix: 统一技能描述读取，修复多行 YAML description 显示为 `|` 及 Gateway API 不读取 SKILL.md 描述
  - 将 read_skill_description() 从 skills.rs 局部函数提升为 blockcell-skills crate 的公共 API
    (openclaw_parser.rs)，使用 serde_yaml 正确解析 YAML frontmatter，修复多行 description
    (description: |) 显示为字面 `|` 的问题
  - Gateway API (/v1/skills, /v1/skills/search) 改用共享 read_skill_description()，支持从
    SKILL.md frontmatter / meta.yaml / meta.json 读取描述
  - 删除 skills.rs 中 70 行重复的 read_skill_description() 实现，消除 slash command 与
    Gateway API 之间的逻辑漂移
  - lib.rs 新增 pub use openclaw_parser::read_skill_description 导出
